refactor: split chat.ts view into smaller components (#36)
Extract three focused modules from the 957-line chat.ts:
- chat-search.ts — search bar UI, text highlighting, match navigation
- chat-attachments.ts — file selection, drag-and-drop, preview management
- chat-messages.ts — message rendering, date separators, status updates

chat.ts remains as the orchestrator wiring these modules together.
All 370 tests pass, tsc clean, no behavior changes.
